# docs/swagger.yaml
basePath: /api
definitions:
  apperror.AppError:
    properties:
      code:
        type: integer
      error_code:
        description: Machine-readable error code
        type: string
      message:
        type: string
    type: object
...
  dto.CreateOrganizationRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      organization_type:
        enum:
        - company
        - store
        type: string
      parent_organization_id:
        type: string
    required:
    - name
    - organization_type
    type: object
...
  dto.JoinOrganizationRequest:
    properties:
      organization_code:
        type: string
    required:
    - organization_code
    type: object
...
  dto.OrganizationResponse:
    properties:
      child_organizations:
        items:
          $ref: '#/definitions/dto.OrganizationResponse'
        type: array
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/dto.UserResponse'
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      member_count:
        type: integer
      name:
        type: string
      organization_type:
        type: string
      parent_organization:
        $ref: '#/definitions/dto.OrganizationResponse'
      parent_organization_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateOrganizationRequest:
    properties:
      description:
        maxLength: 500
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
  dto.UserOrganizationResponse:
    properties:
      is_active:
        type: boolean
      joined_at:
        type: string
      organization:
        $ref: '#/definitions/dto.OrganizationResponse'
      organization_id:
        type: string
      role_in_organization:
        type: string
      user_id:
        type: string
    type: object
  dto.UserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.png
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        type: string
      role:
        example: user
        type: string
      role_id:
        example: b1c2d3e4-f5g6-7890-1234-567890abcdef
        type: string
      username:
        example: johndoe
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Beresin App backend.
  title: Beresin App API
  version: "1.0"
paths:
  /admin/organizations:
    post:
      consumes:
      - application/json
      description: Creates a new organization. Requires platform level permissions.
      parameters:
      - description: Organization Details
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Organization created successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - Admin
      - Organizations
  /admin/organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an organization. Requires platform level permissions.
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Organization deleted successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Admin
      - Organizations
    put:
      consumes:
      - application/json
      description: Updates organization details. Requires admin permissions for the
        organization.
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Organization Details
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Organization updated successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - Admin
      - Organizations
  /admin/organizations/{id}/members:
    get:
      consumes:
      - application/json
      description: Retrieves all members of an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization members retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.UserOrganizationResponse'
            type: array
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get organization members
      tags:
      - Admin
      - Organizations
...
  /organizations:
    get:
      consumes:
      - application/json
      description: Retrieves organizations with optional filters
      parameters:
      - description: Organization Type
        enum:
        - platform
        - company
        - store
        in: query
        name: type
        type: string
      - description: Parent Organization ID
        in: query
        name: parent_id
        type: string
      - description: Is Active
        in: query
        name: active
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Organizations retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.OrganizationResponse'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List organizations
      tags:
      - Organizations
  /organizations/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves organization details by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization retrieved successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get organization by ID
      tags:
      - Organizations
  /organizations/{id}/leave:
    delete:
      consumes:
      - application/json
      description: Allows a user to leave an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully left organization
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Leave organization
      tags:
      - Organizations
      - Users
  /organizations/code/{code}:
    get:
      consumes:
      - application/json
      description: Retrieves organization details by code
      parameters:
      - description: Organization Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization retrieved successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get organization by code
      tags:
      - Organizations
  /organizations/join:
    post:
      consumes:
      - application/json
      description: Allows a user to join an organization using organization code
      parameters:
      - description: Join Organization Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JoinOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully joined organization
          schema:
            $ref: '#/definitions/dto.UserOrganizationResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: User already member of organization
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Join organization
      tags:
      - Organizations
      - Users
...
  /users/me/organizations:
    get:
      consumes:
      - application/json
      description: Retrieves all organizations for the current user
      produces:
      - application/json
      responses:
        "200":
          description: User organizations retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.UserOrganizationResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get user organizations
      tags:
      - Organizations
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"